Google Cloud API added / Redis and Bull queue added / Changed the structure of config : logger location changed
