+++ server/.env_sample
@@ -28,3 +28,6 @@ CLIENT_SECRET=
 REDIRECT_URI=
 REFRESH_TOKEN= 
 TARGET_FOLDER=
+
+#Redis
+redisURI =
