+++ server/.gitignore
@@ -4,6 +4,8 @@ MISC
 Log
 temp
 temp_dbbackupPrac
+util/obsolete
 #heroku doesn't need below
 yarn.lock
 package-lock.json
+
