+++ server/models/index.js
@@ -6,9 +6,10 @@ const path = require('path');
 const Sequelize = require('sequelize');
 const basename = path.basename(__filename);
 const env = process.env.NODE_ENV || 'development';
-const config = require(__dirname + '/../config/config.js')[env];
+const config = require(__dirname + '/../config/seqConfig.js')[env];
 const db = {};
-
+// console.log("[index.js config check] " , require(__dirname + '/../config/config.js'), `  ${env}`)
+// console.log("[index.js config check] " , JSON.stringify(config))
 // don't want to pluralize table name
 config.define = {"freezeTableName" : true}
 config.dialectOptions = {ssl: {
