+++ server/package.json
@@ -8,9 +8,11 @@
     "body-parser": "^1.20.0",
     "bull": "^4.8.3",
     "cors": "^2.8.5",
+    "csv": "^6.1.2",
     "dotenv": "^16.0.0",
     "express": "^4.17.3",
     "googleapis": "^101.0.0",
+    "gzipme": "^1.0.0",
     "mysql2": "^2.3.3",
     "node-cron": "^3.0.0",
     "nodemon": "^2.0.15",
