--- server/util/DBRestoreFromBackup/CSVBackup/company.csv
@@ -1,39 +1,0 @@
-"uuid","companyname","description","benefits","createdAt","updatedAt","deletedAt","job_scraper"
-"3a4c9858-4028-4218-8e9e-e36cc4111290","Abbott",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"c7825ccd-8b9c-4186-b1bb-2fe2d7ba0a80","Aetna, A CVS Company",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"e661a917-a366-4918-a9e8-d29f62bac5e6","Amazon",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"9c02fde6-4988-4f6f-9d5f-a0e56a162e46","American Family Insurance",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"1b217487-9c6b-4921-a712-dbea6754d0c6","Atticus",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"10c4e9dc-588c-4721-b36a-d6269fa2379e","Carvana",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"9a21ebd1-d9e5-416e-9f70-a9d067d9ae53","Chegg",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"ef2496a3-e917-4e6d-9e85-b189b4d5c0dd","Chipotle",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"6a864ab5-73a8-4fd0-9044-5ac9ca30256b","ChowNow",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"cf84302c-9cb8-4a0b-8eb7-ef19d696621f","Commonbond",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"3179bf18-b0a0-4285-962f-feca8e228782","Crew Carwash",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"aa5b4f15-0194-4603-806f-b0e739291b0e","CSAA Insurance Group",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"b460b6d9-14b0-46c3-b81a-fb059e47d8dc","Easterseals NH",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"9ab8c498-d718-4c6a-b36f-8329416c0987","Estee Lauder",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"b1579fd1-c72d-4322-b821-1839e64590b3","Fidelity Investments",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"490a6e3e-a7ee-453f-bc27-e47b9608eb3f","First Republic Bank",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"a1348534-beca-415a-b600-258e311c9b68","Freddie Mac",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"f3f5e51d-feda-42af-a114-34d360267023","Hulu",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"3b59484d-2121-4599-9b44-045967a40a59","LendEDU",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"2ea26dec-5067-49b0-95b8-c93c9ec535d1","Live Nation Entertainment",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"2ec38bb4-48ae-48c7-8244-70a21ed6b4ca","Lowe's",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"f54c4cff-b0f1-46df-a7e5-486f39ef834c","Natixis",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"da2c21a8-c60b-4d62-8382-ead286b2b440","New York Life Insurance Company",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"0221b6d3-cbe1-40bb-b8e3-d8f53c1ceb76","Noodles & Company",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"fb0f2288-0ae9-4aac-a2c4-14be0a214b3a","NVIDIA",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"b2fdb62d-d8ac-49e2-8989-55f8c9079074","Papa John's",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"f28216c0-c922-4da3-947d-42fc473dbd6b","Peloton",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"6e2480e7-322a-42f7-8160-7a972c6f1299","Penguin Random House",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"dbb455d4-ce9b-4f4f-b5f4-b71ec97ed3d8","Pure Insurance",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"efd01b7f-eeef-473f-8764-79abd49dac49","PwC",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"7b10c203-2f2e-4e64-bee3-4283610e88e5","SoFi",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"063dc58d-3d94-447a-a363-4d3d6d483338","Sotheby's",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"3b95c92b-454c-43da-904f-3cceefde4b8c","Target",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"53e89fea-cbad-4c33-826e-bcd65a3ddba0","Teachable",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"e5d27eb4-9527-432d-a4d0-8ea2cfedfeaa","Terminix",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"e2170d3b-e8b4-4380-8dd0-15d31c428f43","Unqork",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"c2c7abab-6d3b-44a7-8710-899889294549","Walmart",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
-"13dc7766-163f-49a5-8e60-3edb6a096377","Weedmaps",NULL,NULL,"2022-05-27 22:45:03.246+00","2022-05-27 22:45:03.246+00",NULL,True
