--- server/util/cronScheduler.js
@@ -1,23 +1,0 @@
-const cron = require('node-cron')
-const {pullJobPostings} = require('../controllers/jobPostingFetcher')
-const {jobPostingDataPurge} = require('../controllers/jobPostingDataPurge')
-
-function dailyJobScraping() {
-    // runs daily at certain hour and minute
-    cron.schedule('48 19 * * *', () => {
-        console.log("[daily job scraping] start! ")
-
-        pullJobPostings().then(async()=>{
-            await jobPostingDataPurge()
-          }).then(()=> {
-            console.log("[daily job scraping] finished! ")
-          })
-    },
-    {
-        // recover missed executions 
-        // recoverMissedExecutions : true
-    })
-}
-
-
-module.exports = {dailyJobScraping}
