+++ server/util/setupDatabase.js
@@ -2,7 +2,7 @@
 const {sequelize} = require('../models')
 const {Client} = require('pg');
 require('dotenv').config()
-const logger = require('./logger')
+const logger = require('../config/logger')
 // const Client= require('pg-promise')
 
 let initialDBconnectionParam =
